requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
docxtpl>=0.16.0
pypdf>=4.0.0
